fix(exchange): stabilize integration tests and adapters (v3.79.6-rc5)
- Tests: quantity decimals may be 0 for contract lots (assert < 0 only)
- Kraken: KuCoin-style display names, valid base64 test secret, ContractInfo numeric fields
- KuCoin/MEXC/Phemex/Huobi/CoinEx/Bitfinex/Deribit: names, JSON, defensive merges, optional Deribit auth

Made-with: Cursor

+++ exchange/kraken/client.go
@@ -426,8 +426,8 @@ type ContractInfo struct {
 	Underlying      string  `json:"underlying"`
 	TickSize        float64 `json:"tickSize"`
 	ContractSize    float64 `json:"contractSize"`
-	MaxPositionSize int     `json:"maxPositionSize"`
-	ImpactMidSize   int     `json:"impactMidSize"`
+	MaxPositionSize float64 `json:"maxPositionSize"`
+	ImpactMidSize   float64 `json:"impactMidSize"`
 	OpeningDate     string  `json:"openingDate"`
 	MarginLevels    []struct {
 		Contracts         int     `json:"contracts"`
